requirements.txt
streamlit
torch
torchvision
pandas
Pillow
plotly
